rename classes and files
