+++ packages/shared_preferences/shared_preferences_foundation/example/lib/main.dart
@@ -34,8 +34,8 @@ class SharedPreferencesDemo extends StatefulWidget {
 class SharedPreferencesDemoState extends State<SharedPreferencesDemo> {
   final SharedPreferencesAsyncPlatform? _prefs =
       SharedPreferencesAsyncPlatform.instance;
-  SharedPreferencesFoundationOptions options =
-      SharedPreferencesFoundationOptions();
+  SharedPreferencesAsyncFoundationOptions options =
+      SharedPreferencesAsyncFoundationOptions();
   static const String _counterKey = 'counter';
   late Future<int> _counter;
